Move server tests to src/server/test/
