+++ src/server/test/server-info-endpoint.js
