+++ src/server/test/sync-endpoint.js
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict'
 import test from 'node:test'
-import { projectKeyToPublicId } from '../../src/utils.js'
+import { projectKeyToPublicId } from '../../utils.js'
 import { createTestServer, randomProjectKeys } from './test-helpers.js'
 
 test('sync endpoint is available after adding a project', async (t) => {
